requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
pytest>=7.4.0
scipy>=1.10.0
